requirements.txt
asgiref==3.8.1
Django==4.2.15
djangorestframework==3.15.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
sqlparse==0.5.1
typing-extensions==4.12.2
